backend/requirements.txt
# Enterprise AI Support Platform - Backend Dependencies
# Python 3.11+

# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.109.1
uvicorn[standard]==0.27.0
python-multipart==0.0.9           # File upload support

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.25
psycopg2-binary==2.9.9            # PostgreSQL driver
alembic==1.13.1                   # Database migrations

# ── Configuration ─────────────────────────────────────────────────────────────
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# ── Authentication ────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0  # JWT
passlib[bcrypt]==1.7.4            # Password hashing

# ── AI / Embeddings ───────────────────────────────────────────────────────────
sentence-transformers==2.4.0      # Semantic embeddings (all-MiniLM-L6-v2)
faiss-cpu==1.7.4                  # Vector similarity search
numpy==1.26.3

# ── LLM ───────────────────────────────────────────────────────────────────────
openai==1.10.0                    # OpenAI GPT API

# ── Document Processing ───────────────────────────────────────────────────────
pymupdf==1.23.8                   # PDF text extraction (fitz)
python-docx==1.1.0                # DOCX text extraction

# ── HTTP / Async ──────────────────────────────────────────────────────────────
httpx==0.26.0
aiofiles==23.2.1

# ── Logging & Monitoring ──────────────────────────────────────────────────────
loguru==0.7.2

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0                     # HTTP client for testing FastAPI
